readded warnings :P

+++ P2/codes/CSC501_KV-master/benchmark/validate.c
@@ -5,13 +5,10 @@
 #include <keyvalue.h>
 #include <fcntl.h>
 
-// added the following libraries
-#include <string.h>
-
 int main(int argc, char *argv[])
 {
     int i=0,number_of_threads = 1, number_of_keys=1024; 
-    unsigned long long tid;
+    int tid;
     __u64 size;
     __u64 key;
     char data[4096],op;
@@ -31,7 +28,7 @@ int main(int argc, char *argv[])
     }
     // Replay the log
     // Validate
-    while(scanf("%c %llu %llu %llu %s",&op, &tid, &key, &size, &data)!=EOF)
+    while(scanf("%c %llu %llu %d %s",&op, &tid, &key, &size, &data)!=EOF)
     {
         if(op == 'S')
         {
